requirements.txt
fastapi
uvicorn[standard]
sqlmodel
# Database drivers (Async)
aiosqlite 
# AI & Vector DB
qdrant-client
google-generativeai
sentence-transformers
# Utilities
python-dotenv
pydantic-settings
httpx